pyproject.toml
[tool.poetry]
name = "ppsc-paper-bank"
version = "0.1.0"
description = "A FastAPI-based MCQ power bank system"
authors = ["User"]
readme = "README.md"
packages = [
    { include = "app" },
    { include = "main.py" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
sqlmodel = "^0.0.24"
python-dotenv = "^1.1.0"
uvicorn = "^0.34.2"
# MySQL drivers
pymysql = "^1.1.1"
mysql-connector-python = "^9.0.0"
cryptography = "^3.4.8"
# Web scraping
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
lxml = "^5.4.0"
# Template and PDF generation
jinja2 = "^3.1.6"
xhtml2pdf = "^0.2.17"
# weasyprint = "^65.1"  # Optional - requires GTK libraries on Windows
# Other utilities
asgiref = "^3.8.1"
httpx = "^0.28.1"
mcp = "^1.23.3"
markdownify = "^1.2.2"
openai-agents = {extras = ["litellm", "sqlalchemy"], version = "^0.6.2"}
aiomysql = "^0.3.2"
pyodbc = "^5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.23.5"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
